+++ lib/operations/components/callback.rb
@@ -2,11 +2,13 @@
 
 require "operations/components/base"
 
-# This base component handles `on_failure:` and `on_success` callbacks
+# This base component handles `on_failure:` and `on_success:` callbacks
 # passed to the command. Every callback entry is called outside of the
 # operation transaction and any exception is rescued here so the result
 # of the whole operation is not affected. Additionally, any callback
 # failures will be reported with the command error reporter.
+# The original operation result will be optionally passed as the second
+# positional argument for the `call` method.
 class Operations::Components::Callback < Operations::Components::Base
   include Dry::Monads::Do.for(:call_entry)
 
@@ -14,8 +16,8 @@ class Operations::Components::Callback < Operations::Components::Base
 
   private
 
-  def call_entry(entry, params, **context)
-    result = yield(entry_result(entry, params, **context))
+  def call_entry(entry, operation_result, **context)
+    result = yield(entry_result(entry, operation_result, **context))
 
     Success(result)
   rescue Dry::Monads::Do::Halt => e
@@ -24,11 +26,14 @@ class Operations::Components::Callback < Operations::Components::Base
     Failure(e)
   end
 
-  def entry_result(entry, params, **context)
+  def entry_result(entry, operation_result, **context)
     args = call_args(entry, types: %i[req opt])
 
-    if args.one?
-      entry.call(params, **context)
+    case args.size
+    when 1
+      entry.call(operation_result.params, **context)
+    when 2
+      entry.call(operation_result.params, operation_result, **context)
     else
       entry.call(**context)
     end